+++ src/defs.rs
@@ -15,11 +15,11 @@ pub const BLACK: Side = 1;
 pub const FEN_START_POSITION: &str = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1";
 
 pub const NR_OF_SQUARES: usize = 64;
-pub const NR_OF_FILES: u8 = 8;
-pub const NR_OF_RANKS: u8 = 8;
-pub const NR_OF_PIECES: u8 = 6;
-pub const NR_OF_CASTLING_PERMISSIONS: u8 = 16; // 0-15
-pub const EACH_SIDE: u8 = 2;
+pub const NR_OF_FILES: usize = 8;
+pub const NR_OF_RANKS: usize = 8;
+pub const NR_OF_PIECES: usize = 6;
+pub const NR_OF_CASTLING_PERMISSIONS: usize = 16; // 0-15
+pub const EACH_SIDE: usize = 2;
 
 pub struct Castling;
 impl Castling {
